Open source publish: 87a2838e..c9a12cf6
X-Publish-Start-Commit: 87a2838e5e0d3ae5d62a59be514a030debf50caa
X-Publish-End-Commit: c9a12cf60f36541a4f6210b10c307e2fe423b19b

---

Cast to unsigned int 32

---

Changes required to eventually enable pedantic flag

---

Explicitly install clang 6.0

---

Allow zero observations during RTCM -> SBP conversion

Version 5.8.14 of the TeseoV firmware introduced a configuration option which allows MSM7 messages to be sent when no satellites are available. Enabling this setting allows the possibility of outputting MSG_GPS_TIME_GNSS messages from the Starling GNSS engine in situations where no satellites are in view (or the antenna is disconnected). This is required in order to allow synchronisation of external sensor data with Starling time during GNSS signal blockage events

---

rust/build.rs: Always invoke cmake

Build may fail locally on macOS without this change.

---

rust/lib.rs: Fix fetch_io argument parsing

* rust/build.rs: Always invoke cmake

Build may fail locally on macOS without this change.

* rust/lib.rs: Fix fetch_io argument parsing

fetch_io had an impossible condition to show the help,
and it is also not desirable as the commands can be used
without any file arguments.

Additionally only consider the last two cli args, and check
that the input file exists before assuming the args are
intended to be file args, thereby allowing '-x ..' style
cli arguments to be passed through to the C.

* rtcm3tosbp.c: Allow use of -x=...

This allows slightly more readable clis, without resorting
to getopt_long.

---

system-timing: Define API

---

system-timing: time truth unit test

---

system-timing: time truth implementation

---

system-timing: rtcm2sbp unit tests

---

system-timing: updated time truth for gps time

---

system-timing: rtcm2sbp converter

---

system-timing: update utilities

---

system-timing: remove RTCM In Converter warning

---

system-timing: refactor time truth to update from sbp

---

* Add option to force output COG field in VTG sentence

It allows a configurable speed threshold (in m/s) to be defined
to control the lowest speed at which Course Over Ground messages
within the VTG, RMC and PUBX sentences will be emitted.

* Only recalculate COG if SOG surpasses threshold

When the cog_threshold_mps is used to always emit COG even at low speed,
the COG can be very sensitive to small amounts of noise in the velocity.

This changes add a way of specifying a threshold for being stationary.
When the speed is below this threshold, new COGs will not be computed
and instead the previously computed COG will be instead emitted.

* Fix function comments

---

Move c++ guard

---

Update cog_stationary_threshold_mps name and set new default

* Update cog_stationary_threshold_mps to cog_update_threshold_mps

* Change cog_update_threshold_mps default to 0.1

This is a better "reasonable default" as will work better when the
cog_threshold_mps is set to 0.

* Fix unknown type name 'size_t'

---

Allow decoding of logs which contain rover observations but no ephemerides

The recently introduced "time truth" concept assumes that ephemerides
are always available, but this is not always the case (e.g. RTCM logs
from the USB stick of a STEP with F9P). The current implementation
causes all GLO observations to be unexpectedly dropped from these kind
of logs since the checks against GLO_SANITY_THRESHOLD_S always fail in
add_glo_obs_to_buffer().

---

Add preliminary support for ISA100C imu

* Fix rate for ISA100C: according to
https://docs.novatel.com/oem7/Content/SPAN_Logs/RAWIMUS.htm it should be
200Hz.  Also add rates and scaling factors for all other known IMU
types.

* Add preliminary support for sending MSG_IMU_RAW and MSG_IMU_AUX messages

* Emulate configuration/orientation of an ASM-330 installed in a STEP

* Address 2 of the points raised by @christian: Use ldexp() and native IMU reference frame

* Address the other two points raised by @christian:
1. attempt to dynamically compute the IMU data raet
2. rate limit MSG_IMU_AUX messages to 1Hz

* Add handling for end-of-week rollover

---

Don't misuse SBP obs header n_obs field

---

Add SonarCloud CI

---

Resolve SonarCloud "quality gate" issues

---

system-timing: remove GLONASS observation time threshold

---

add unit test for invalid pseudorange

---

SonarCloud CI: Add checkout depth 0

---

Fuzz testing

---

SonarCloud CI: Disable multithreaded analysis and analysis cache settings

---

* NDF SBAS decoding

* Fix buffer overrun in sbas_decode_subframe for F9 series

* Update decoding of BDS SSR orbit messages to latest RTCM draft

* Use HNR PVT to produce high rate heading

Replaces ORIENT_EULER output from NAV ATT parsing
while still using all message values from NAV ATT
except for heading, heading_acc and tow.

* Mitigate WN issue for MSG_GNSS_TIME_OFFSET

* Fix test case now that the geoid height fix has been merged

* increase SBAS PRN numbers by one

---

+++ c/src/gnss_converters_extra/sbp_conv.c
@@ -116,5 +116,5 @@ size_t sbp_conv(sbp_conv_t conv,
     }
     default: { break; }
   }
-  return fifo_read(&conv->fifo, wbuf, (u8)wlen);
+  return fifo_read(&conv->fifo, wbuf, (u32)wlen);
 }
